requirements.txt
requests==2.31.0
python-dotenv==1.0.1
numpy==1.26.4
scikit-learn==1.4.1.post1
